Improve plasticity performance by 15%
Synapses have to be updated under two conditions: When their source
neuron expires or when it spikes:

1. Expiration: It is guaranteed that the neuron did not spike since its
last update. If it did it woul've been updated previously and not be
spiking now. Thus 'pre' is always false.
2. Spike: It is guaranteed that the neuron did not spike since its last
update. (same reasoning). Thus 'pre' is false for all but the very last
update step which was triggered by the very spike we are processing.

Hence we only need to query the history for 'post' events.