Implement synthetic model for testing purposes
Implement a synthetic model with a connectivity of 0.1% and a
connectivity of 1%. Spikes result in a simple atomic increment of the
neuron potential. As the synapse count varies from 100M to 1B, the
neuron count grows from 300k to 1M. The runtime grows quadratically as
opposed to linearly because we now spill out of cache.